[REF] Use dtype argument of `conv_index_pattern`

+++ einconv/einconvnd.py
@@ -385,7 +385,8 @@ def _conv_einsum_operands(
             padding=padding if isinstance(padding, str) else t_padding[n],
             dilation=t_dilation[n],
             device=input.device,
-        ).to(input.dtype)
+            dtype=input.dtype,
+        )
         for n in range(N)
     )
