+++ einconv/unfoldnd.py
@@ -76,7 +76,8 @@ def unfoldNd(
             padding=t_padding[n],
             dilation=t_dilation[n],
             device=input.device,
-        ).to(input.dtype)
+            dtype=input.dtype,
+        )
         for n in range(N)
     ]
